Add files via upload
The bot implementation work as follows- the user will provide his weight and height as input, according to which the bot will respond. List trainer package is used here, due to which the chatbot will be able to recognize abbreviations as "u"- “you" etc. The code calculates the BMI and range the person as per his weight and provide the user with appropriate calorie intake as per his weight further it gives information about exercises and diet plan for weight maintenance.